# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "properties": {
          "code": {
            "description": "错误代码",
            "title": "Code",
            "type": "string"
          },
          "message": {
            "description": "错误说明",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReportExportRequest": {
        "properties": {
          "benchmark_task_id": {
            "title": "Benchmark Task Id",
            "type": "string"
          }
        },
        "required": [
          "benchmark_task_id"
        ],
        "title": "ReportExportRequest",
        "type": "object"
      },
      "ReportListPayload": {
        "properties": {
          "reports": {
            "items": {
              "$ref": "#/components/schemas/ReportPayload"
            },
            "title": "Reports",
            "type": "array"
          }
        },
        "required": [
          "reports"
        ],
        "title": "ReportListPayload",
        "type": "object"
      },
      "ReportListResponse": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReportListPayload"
              },
              {
                "type": "null"
              }
            ],
            "description": "业务数据"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorBody"
              },
              {
                "type": "null"
              }
            ],
            "description": "错误信息"
          },
          "success": {
            "default": true,
            "description": "请求是否成功",
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "ReportListResponse",
        "type": "object"
      },
      "ReportPayload": {
        "properties": {
          "artifact_dir": {
            "title": "Artifact Dir",
            "type": "string"
          },
          "benchmark_definition_id": {
            "title": "Benchmark Definition Id",
            "type": "string"
          },
          "benchmark_task_id": {
            "title": "Benchmark Task Id",
            "type": "string"
          },
          "created_at_utc": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At Utc"
          },
          "dut_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dut Model"
          },
          "json_path": {
            "title": "Json Path",
            "type": "string"
          },
          "markdown_path": {
            "title": "Markdown Path",
            "type": "string"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "report_id": {
            "title": "Report Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "$ref": "#/components/schemas/BenchmarkTaskSummaryPayload"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at_utc": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At Utc"
          }
        },
        "required": [
          "report_id",
          "benchmark_task_id",
          "project_id",
          "benchmark_definition_id",
          "dut_model",
          "title",
          "status",
          "artifact_dir",
          "json_path",
          "markdown_path",
          "summary",
          "created_at_utc",
          "updated_at_utc"
        ],
        "title": "ReportPayload",
        "type": "object"
      },
      "ReportResponse": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReportPayload"
              },
              {
                "type": "null"
              }
            ],
            "description": "业务数据"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorBody"
              },
              {
                "type": "null"
              }
            ],
            "description": "错误信息"
          },
          "success": {
            "default": true,
            "description": "请求是否成功",
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "ReportResponse",
        "type": "object"
      },
...
      "RunResponse": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunPayload"
              },
              {
                "type": "null"
              }
            ],
            "description": "业务数据"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorBody"
              },
              {
                "type": "null"
              }
            ],
            "description": "错误信息"
          },
          "success": {
            "default": true,
            "description": "请求是否成功",
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "RunResponse",
        "type": "object"
      },
...
      "ScenarioLaunchRequest": {
        "properties": {
          "auto_start": {
            "default": true,
            "title": "Auto Start",
            "type": "boolean"
          },
          "map_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Map Name"
          },
          "metadata": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScenarioMetadataPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "scenario_id": {
            "title": "Scenario Id",
            "type": "string"
          },
          "sensor_profile_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sensor Profile Name"
          },
          "template_params": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                },
                {
                  "type": "string"
                }
              ]
            },
            "title": "Template Params",
            "type": "object"
          },
          "timeout_seconds": {
            "anyOf": [
              {
                "maximum": 86400.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout Seconds"
          },
          "traffic": {
            "$ref": "#/components/schemas/TrafficPayload"
          },
          "weather": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WeatherPayload"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "scenario_id"
        ],
        "title": "ScenarioLaunchRequest",
        "type": "object"
      },
      "ScenarioMetadataPayload": {
        "additionalProperties": true,
        "properties": {
          "author": {
            "default": "",
            "title": "Author",
            "type": "string"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "dut_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dut Model"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          }
        },
        "title": "ScenarioMetadataPayload",
        "type": "object"
      },
...
      "SensorProfileResponse": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SensorProfilePayload"
              },
              {
                "type": "null"
              }
            ],
            "description": "业务数据"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorBody"
              },
              {
                "type": "null"
              }
            ],
            "description": "错误信息"
          },
          "success": {
            "default": true,
            "description": "请求是否成功",
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "SensorProfileResponse",
        "type": "object"
      },
      "SensorProfileSaveRequest": {
        "properties": {
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "profile_name": {
            "title": "Profile Name",
            "type": "string"
          },
          "sensors": {
            "items": {
              "$ref": "#/components/schemas/SensorSpecPayload"
            },
            "title": "Sensors",
            "type": "array"
          },
          "vehicle_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vehicle Model"
          }
        },
        "required": [
          "profile_name",
          "display_name",
          "sensors"
        ],
        "title": "SensorProfileSaveRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "SimChip Nexus 芯境智测平台提供 CARLA 场景仿真、硬件在环测试、设备接入和评测报告服务。\n\n- `/docs`：接口文档\n- `/`：平台展示页\n- `/ui`：Web 测评控制台",
    "title": "SimChip Nexus API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/reports": {
      "get": {
        "operationId": "list_reports_reports_get",
        "parameters": [
          {
            "in": "query",
            "name": "benchmark_task_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Benchmark Task Id"
            }
          },
          {
            "in": "query",
            "name": "project_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "查询报告列表",
        "tags": [
          "报告中心"
        ]
      }
    },
    "/reports/export": {
      "post": {
        "operationId": "export_report_reports_export_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "导出基准任务报告",
        "tags": [
          "报告中心"
        ]
      }
    },
...
    "/reports/{report_id}": {
      "get": {
        "operationId": "get_report_reports__report_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "title": "Report Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "查询单个报告",
        "tags": [
          "报告中心"
        ]
      }
    },
...
    "/scenarios/launch": {
      "post": {
        "description": "前端只提交场景、地图、天气、传感器和背景交通参数。后端生成 per-run 的 native runtime 输入并创建 run，可按需自动加入执行队列。",
        "operationId": "launch_scenario_scenarios_launch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioLaunchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "按场景启动配置创建运行",
        "tags": [
          "场景管理"
        ]
      }
    },
...
    "/scenarios/sensor-profiles/{profile_name}": {
      "put": {
        "description": "用于运维界面维护与车型绑定的 YAML 传感器模板。场景运行仍只引用 profile_name。",
        "operationId": "save_sensor_profile_endpoint_scenarios_sensor_profiles__profile_name__put",
        "parameters": [
          {
            "in": "path",
            "name": "profile_name",
            "required": true,
            "schema": {
              "title": "Profile Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SensorProfileSaveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SensorProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "创建或更新传感器配置模板",
        "tags": [
          "场景管理"
        ]
      }
    },
...
  }
}